requirements.txt
addict==2.4.0
einops==0.3.2
inplace_abn==1.1.1.dev3+ga8a3f82
matplotlib==3.5.0
Mosek==9.2.3
numpy==1.20.3
pandas==1.3.3
Pillow==9.1.1
pycocotools==2.0.4
PyYAML==6.0
randaugment==1.0.2
scikit_learn==1.1.1
scipy==1.7.1
seaborn==0.11.2
termcolor==1.1.0
timm==0.4.12
torch==1.8.2+cu111
torchvision==0.9.2+cu111
tqdm==4.62.3
yapf==0.31.0
